+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v4.0.0 (by Gecko) - ??/??/????
 - `ViewerGetPickedObjectsList` now requires user-provied PickList
 - New constants for `ViewerSetAntiAliasing`: `aa6x = 6`, `aa8x = 7`, `aa16x = 8`, `csa8x = 9`, `csa8xHQ = 10`, `csa16x = 11`, `csa16xHQ = 12`
 - `CameraZoomAll` now requires a Viewer as a second parameter
+- `SpriteNoZWrite` was removed, use `MaterialSetDepthWrite` instead
 - `MaterialCullFrontFaces` was removed
 - `MaterialSetZWrite` is now `MaterialSetDepthWrite`
 - New function `MaterialSetDepthTest`